griot-knowledge/requirements-docker.txt
# =====================================================
# GriotKnowledge — requirements.txt
# Versions vérifiées et alignées sur votre environnement
# =====================================================

# API & Serveur
fastapi==0.133.1
uvicorn==0.41.0
python-multipart==0.0.22
starlette==0.52.1

# RAG & LLM orchestration
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.16
langchain-huggingface==1.2.0
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph==1.0.9

# Embeddings & Modèles
sentence-transformers==5.2.3
torch==2.10.0
transformers==5.2.0
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==0.33.4

# Base de données vectorielle
qdrant-client==1.17.0

# Validation & Configuration
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1

# Calcul & ML utilitaires
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.1

# HTTP & Réseau
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3

# Utilitaires
tqdm==4.67.3
rich==14.3.3
pyyaml==6.0.3